test; pruebas de integración usando spring boot test, junit y testcontainer

+++ src/main/java/com/unimag/Tienda/Entidad/Producto.java
@@ -19,8 +19,9 @@ public class Producto {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String nombreProducto;
-    private BigDecimal precio;
-    private  int stock;
+    private Double precio;
+    private  Integer stock;
+
 
    @OneToMany(mappedBy = "producto")
    private List<ItemPedido> itemPedidos;
